refactor(domain): DepositHistory 도메인 패키지 이동(attendance -> cohort)
