+++ module-domain/src/main/java/com/longrunpc/domain/cohort/entity/DepositType.java
@@ -1,4 +1,4 @@
-package com.longrunpc.domain.attendance.entity;
+package com.longrunpc.domain.cohort.entity;
 
 public enum DepositType {
     INITIAL,
